[MT-4619] Reorganize gallery components
Lifting their higher in the directories tree
