+++ packages/slate-editor/src/modules/editor-v4-galleries/components/Gallery/index.ts
