+++ packages/slate-editor/src/modules/editor-v4-galleries/components/Gallery/lib/calculateLayout.ts
