--- packages/slate-editor/src/modules/editor-v4-galleries/components/GalleryElement/index.ts
@@ -1,1 +1,0 @@
-export { GalleryElement } from './GalleryElement';
