--- packages/slate-editor/src/modules/editor-v4-galleries/components/GalleryMenu/index.ts
@@ -1,1 +1,0 @@
-export { GalleryMenu } from './GalleryMenu';
