+++ packages/slate-editor/src/modules/editor-v4-galleries/components/index.ts
@@ -1,2 +1,1 @@
 export { GalleryElement } from './GalleryElement';
-export { GalleryMenu } from './GalleryMenu';
